+++ compiler/astdef.nim
@@ -127,6 +127,8 @@ type
     sfWasGenSym       # symbol was 'gensym'ed
     sfForceLift       # variable has to be lifted into closure environment
 
+    sfDeferredAlign   # field has deferred align pragma expression
+
     sfDirty           # template is not hygienic (old styled template) module,
                       # compiled from a dirty-buffer
     sfCustomPragma    # symbol is custom pragma template
@@ -398,6 +400,7 @@ type
     tfSendable
     tfImplicitStatic
     tfDeferredSize      # size pragma has deferred expression
+    tfDeferredAlign     # align pragma has deferred expression
 
   TTypeFlags* = set[TTypeFlag]
 
@@ -710,6 +713,7 @@ type
       guardImpl*: PSym
       bitsizeImpl*: int
       alignmentImpl*: int        # for alignment
+      alignExprImpl*: PNode      # deferred align expression (nil = use alignmentImpl)
     else: nil
     magicImpl*: TMagic
     typImpl*: PType
@@ -798,6 +802,7 @@ type
     alignImpl*: int16             # the type's alignment requirements
     paddingAtEndImpl*: int16      #
     sizeExprImpl*: PNode          # deferred size expression (nil = use sizeImpl)
+    alignExprImpl*: PNode         # deferred align expression (nil = use alignImpl)
     locImpl*: TLoc
     typeInstImpl*: PType          # for generic instantiations the tyGenericInst that led to this
                               # type.